AssetLoader: error checking
If a texture is loaded early during initialization and cannot be found, a default image is used. If the error occurs early enough, the default image isn't there yet.

GVRPhongLayeredShader is only capable of layering two textures. Now we check if the FBX file layers more than two and no longer crash the shader.
This patch checks for these cases.