Cursor controller bug fixes
1. Fix cursor depth scaling in Gear controller
2. Set controller active state in Gaze controller
3. Combined GVRInputManager and GVRInputManagerImpl into one class
4. GVRPicker enables SEND_TO_HIT_OBJECT when using controller
5. Fix Gear controller duplicate motion events
Picking with the controller was wonky because the controller was
automatically generating motion events. GVRViewSceneObject and
GVRWidgetPlugin also convert controller events into motion events within
the quad representing the Android view. This patch suppresses the motion
events automatically generated by the controller which fixes picking
inside Android views. GVRGearController.sendEventsToActivity is added to
enable routing the events to the activity if this is desired.